+++ XprTrueHoleHttpClient.unittests/Properties/AssemblyInfo.cs
