+++ XprTrueHoleHttpClient.unittests/TestTHApiClient.cs
@@ -4,7 +4,7 @@ using System.Net;
 using System.Net.Http;
 using System.Text;
 
-namespace Hypertherm.TrueHoleHttpClient.unittests
+namespace Hypertherm.XprTrueHoleHttpClient.unittests
 {
     [TestFixture]
     public class TestTHApiClient
