+++ XprTrueHoleHttpClient.unittests/packages.config
